Patient profile add notification

+++ database/migrations/2023_07_04_161951_create_patients_table.php
@@ -15,8 +15,8 @@ class CreatePatientsTable extends Migration
     {
         Schema::create('patients', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
-            $table->integer('patient_id');
+            $table->bigInteger('user_id');
+            $table->bigInteger('patient_id');
             $table->string('gender')->nullable();
             $table->string('blood')->nullable();            
             $table->date('dob')->nullable();
